+++ src/components/NftCard/NftCard.scss
@@ -1,55 +1,51 @@
 @import "src/styles/index";
 
 .nft-card {
-    @extend %button;
-    @extend %button-border-style;
-
-    flex-direction: column;
+    display: grid;
+    grid-template-rows: 9.25rem 1fr 1fr 1fr 1fr;
+    align-items: center;
     border-radius: .25rem;
-    padding: 0;
-    text-align: left;
-    overflow: hidden;
 
-    @include for-size(tablet-landscape-up) {
-        border-radius: .5rem;
+    width: 9.25rem;
+    height: 12.25rem;
+    background-color: var(--c-grey);
+
+    @include for-size(desktop-up) {
+        grid-template-rows: 16.25rem 1fr 1fr 1fr 1fr;
+        width: 16.25rem;
+        height: 21.5rem;
     }
 
     &__img {
-        width: 100%;
-        aspect-ratio: 1;
+        height: 100%;
+        width:100%;
         object-fit: contain;
-    }
-
-    &__info-wrapper {
-        width: 100%;
-        padding: .5rem .75rem .75rem;
-        font-size: .875rem;
-        background-color: var(--c-grey);
-
-        @include for-size(desktop-up) {
-            padding: .75rem 1.25rem 1rem;
-            font-size: 1rem;
-        }
-    }
-
-    &__name,
-    &__price {
-        @extend %text-ellipsis;
     }
     
     &__name {
+        grid-row: 3;
+        margin-left: .5rem;
+        font-size: .75rem;
         font-weight: var(--fw-700);
         color: var(--c-white);
+
+        @include for-size(desktop-up) {
+            margin-left: 1.25rem;
+            font-size: 1rem;
+        }
     }
     
     &__price {
-        margin-top: .5rem;
+        grid-row: 4;
+        margin-left: .5rem;
+        font-size: .75rem;
         font-weight: var(--fw-400);
-        font-family: var(--f-mono);
         color: var(--c-light-grey);
+        font-family: var(--f-mono);
 
         @include for-size(desktop-up) {
-            margin-top: .75rem;
+            margin-left: 1.25rem;
+            font-size: 1rem;
         }
     }
 }
